move actionSchema to shared
