+++ front/components/agent_builder/capabilities/mcp/utils/formDefaults.ts
@@ -3,7 +3,7 @@ import set from "lodash/set";
 import type {
   AdditionalConfigurationInBuilderType,
   MCPServerConfigurationType,
-} from "@app/components/agent_builder/AgentBuilderFormContext";
+} from "@app/components/shared/tools_picker/types";
 import { getMCPServerRequirements } from "@app/lib/actions/mcp_internal_actions/input_configuration";
 import type { MCPServerViewType } from "@app/lib/api/mcp";
