+++ front/components/agent_builder/capabilities/mcp/validation/schemaBuilders.ts
@@ -1,5 +1,6 @@
 import { z } from "zod";
 
+import { VALIDATION_MESSAGES } from "@app/components/agent_builder/capabilities/mcp/utils/validationMessages";
 import {
   childAgentIdSchema,
   dataSourceConfigurationSchema,
@@ -10,8 +11,7 @@ import {
   mcpTimeFrameSchema,
   reasoningModelSchema,
   secretNameSchema,
-} from "@app/components/agent_builder/AgentBuilderFormContext";
-import { VALIDATION_MESSAGES } from "@app/components/agent_builder/capabilities/mcp/utils/validationMessages";
+} from "@app/components/shared/tools_picker/types";
 import type { MCPServerRequirements } from "@app/lib/actions/mcp_internal_actions/input_configuration";
 
 /**
