+++ front/components/agent_builder/submitAgentBuilderForm.ts
@@ -1,12 +1,10 @@
-import type {
-  AdditionalConfigurationInBuilderType,
-  AgentBuilderFormData,
-} from "@app/components/agent_builder/AgentBuilderFormContext";
+import type { AgentBuilderFormData } from "@app/components/agent_builder/AgentBuilderFormContext";
 import { DROID_AVATAR_URLS } from "@app/components/agent_builder/settings/avatar_picker/types";
 import {
   expandFoldersToTables,
   getTableIdForContentNode,
 } from "@app/components/agent_builder/shared/tables";
+import type { AdditionalConfigurationInBuilderType } from "@app/components/shared/tools_picker/types";
 import type { TableDataSourceConfiguration } from "@app/lib/api/assistant/configuration/types";
 import { clientFetch } from "@app/lib/egress/client";
 import type { AdditionalConfigurationType } from "@app/lib/models/agent/actions/mcp";
